Implement UserDefinedChargeDataProvider and add ChargeData model

+++ database/migrations/2022_03_06_002104_create_charges_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->id();
             $table->timestamps();
             $table->foreignIdFor(RecurringExpense::class);
-            $table->float("amount", unsigned: true);
+            $table->integer("amount", unsigned: true);
             $table->dateTime("charged_at");
             $table->boolean("draft");
             // TODO: add date for draft?
